file extraction added

+++ app.py
@@ -14,21 +14,24 @@ def resource_path(relative_path):
     return os.path.join(base_path, relative_path)
 
 def delete_dup_file():
-    print("info: Checking duplicate files. ")
-    file_to_del = {}
-    items = os.listdir(from_path)
-    for names in items:
-        if names.startswith("ItemReport"):
-            full_file_path = os.path.join(from_path, names)
-            file_to_del[names] = full_file_path
-        # else:
-        #     print("info: no duplicate files found")
-    # print("Deleting: Some file with the same name is already there")
-    for name,path in file_to_del.items():
-            print("Deleting: "+name+" file with the same name is already there")
-            os.remove(path)
-            print("success: Done Deleting"+name)
-
+    prompt = input("prompt: You wanted to Delete all files with the name ItemReport in Downloads:(y/n) ")
+    if(prompt == "y" or prompt == "Y"):
+        print("info: Checking duplicate files. ")
+        file_to_del = {}
+        items = os.listdir(from_path)
+        for names in items:
+            if names.startswith("ItemReport"):
+                full_file_path = os.path.join(from_path, names)
+                file_to_del[names] = full_file_path
+            # else:
+            #     print("info: no duplicate files found")
+        # print("Deleting: Some file with the same name is already there")
+        for name,path in file_to_del.items():
+                print("Deleting: "+name+" file with the same name is already there")
+                os.remove(path)
+                print("success: Done Deleting"+name)
+    else:
+        print("info: Ignoring duplicate files in downloads")
 
 def find_file():
     while(True):
@@ -116,7 +119,7 @@ if(os.path.exists(from_path)):
     else:
         from_path = from_path+"\\"+report_fname
 shutil.move(from_path,des_path)
-print("\nSuccess: moving file.")
+print("\nSuccess: moved file.")
 
 
 driver.close()
@@ -126,6 +129,7 @@ time.sleep(3)
 
 with zipfile.ZipFile(os.path.join(des_path,report_fname),"r") as zip_ref:
     zip_ref.extractall("zip")
+print("info: zip file extracted")
 
 
 print("\n\nComplete....")
